+++ NEWS.md
@@ -1,4 +1,8 @@
-# vaster 0.0.0.9002
+# vaster dev
+
+* New function `vaster_long()` to convert extent, dimension, and optionally data to long form XYZ in raster or trad matrix order. 
+
+* All obvious raster functions now included (needs more testing). 
 
 * Version bump to allow {grout} to depend on vaster for cell and extent ops.